feat(infra): trying to move the gitcoin.ts to a shared lib file ...
